move things around as per clean arch
