+++ src/FamilyHubs.ServiceDirectory.Infrastructure/FamilyHubs.ServiceDirectory.Infrastructure.csproj
@@ -30,4 +30,8 @@
 	  </PackageReference>
 	</ItemGroup>
 
+	<ItemGroup>
+	  <ProjectReference Include="..\FamilyHubs.ServiceDirectory.Core\FamilyHubs.ServiceDirectory.Core.csproj" />
+	</ItemGroup>
+
 </Project>
