+++ src/FamilyHubs.ServiceDirectory.Web/Pages/PostcodeSearch.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@using FamilyHubs.ServiceDirectory.Infrastructure.Services.Interfaces
+@using FamilyHubs.ServiceDirectory.Core.Postcode.Model
 @model FamilyHubs.ServiceDirectory.Web.Pages.PostcodeSearchModel
 @{
     ViewData["Title"] = "Search for local family support and services";
